vcs_info: choose the nearest repository
Previously vcs_info didn't handle nested repositories of different types
correctly: for example, when in a hg repository, that is cloned inside of a git
working tree (not as a submodule), it was still showing git branch.

Choose the repository that's nearest to the current working directory. And
since base directories are always parents of the current directory - just
choose the longest path.